Replaced hardcoded path to .vim directory
In Grammarian plugin's code the vim's home directory was hard coded
with an absolute path to the creator of the code. This was changed
to an expanding variable that hopefully works across platforms.

Also the recompilation of error files was resulting in errors on my
computer. It found the files 'unreadable', but when writing them,
couldn't overwrite, so a '!' was added to force the recompilation
i.o.w. enabling overwriting.